task_manager/requirements.txt
Django==4.x.x
djangorestframework==3.x.x
django-filter==2.x.x